ci(types): pyright in basic mode, 0 errors — narrow `c` to `ScalarCurvature`
`typeCheckingMode` goes from "off" to "basic". Basic mode reported 110 errors
(reportOperatorIssue 49, reportArgumentType 46, reportReturnType 5, 10 others);
the tree is now clean with no rule disabled, no `# type: ignore` added, and
`reportUndefinedVariable` still at "error".

93 of the 110 were one thing: `Curvature = ScalarCurvature |
Sequence[ScalarCurvature]` was the annotation on every manifold's `c`, so
`jnp.sqrt(c)`, `c * x` and `1/c` were all checked against a sequence that
supports none of them. Every single manifold and the `Manifold` protocol itself
now declare `c: ScalarCurvature` (286 annotations, 19 modules); only
`ProductManifold` keeps the union, and it still satisfies the protocol because
an implementation's parameter may be wider than the one it implements.
`ScalarCurvature` is exported from `hyperbolix.manifolds` next to `Curvature`.

Three real defects fell out of the exercise:

* the `Manifold` protocol was missing the trailing `version_idx` on `dist` /
  `dist_0` and `atol` on `is_in_manifold` / `is_in_tangent_space` that its own
  docstring promises, so `compute_pairwise_distances` did not type-check
  against it;
* `hyperbolix.utils.__all__` listed `MATMUL_PRECISION`, which moved to
  `hyperbolix.utils.precision` and was never re-imported, so
  `from hyperbolix.utils import *` raised `AttributeError`;
* `hyp_pad2d(pad_mode="origin")` documents `c` as required but never checked
  it, failing with an opaque `TypeError: unsupported operand type(s) for /`.
  It now raises a `ValueError` naming the argument.

The rest: `scalar_mul`'s `r` widened from `float` to `float | Float[Array, ""]`
(what GyroBatchNorm/GyroRMSNorm already pass, and what
`ProperVelocity._scalar_mul` already declared); `float()` around two
`(x) ** 0.5` init bounds, because `float.__pow__` is typed as returning `Any`
and blocked the `float | None` narrowing; `cast` at four points where a
runtime-validated concrete manifold is reached through the protocol; and
annotations on six locals whose type `@jax.jit` erases (jax types the decorator
as returning an untyped callable).

Annotations and narrowing only. No numerical change: the two `float()` calls
are bit-pattern identities, the `cast`s are runtime no-ops, and the new
`ValueError` only fires where the old code already raised.

+++ hyperbolix/nn_layers/hyperboloid_conv.py
@@ -851,7 +851,9 @@ class HypConv2DHyperboloidILNN(nnx.Module):
         out_spatial = out_channels - 1
         logcat_spatial = logcat_out_ambient_dim - 1
         if kernel_init_std is None:
-            kernel_init_std = (1.0 / out_spatial) ** 0.5
+            # float(): `float.__pow__` is typed as returning Any, which would leave
+            # `kernel_init_std` at its declared `float | None` for the type checker.
+            kernel_init_std = float((1.0 / out_spatial) ** 0.5)
         kernel_init = kernel_init_std * jax.random.normal(rngs.params(), (out_spatial, logcat_spatial), dtype=param_dtype)
         self.kernel = nnx.Param(kernel_init)
         self.bias = nnx.Param(jnp.zeros((out_spatial, 1), dtype=param_dtype))
